backend/requirements.txt
asgiref==3.8.1
Django==5.1.1
sqlparse==0.5.1
djangorestframework
djangorestframework-simplejwt
django-cors-headers
